Cambiados los tipos de vars para ahorrar mem

+++ src/ejercicios/Ejercicio2.java
@@ -4,7 +4,7 @@ public class Ejercicio2 {
     public static void main(String[] args) {
         // ? Programa que con un bucle for muestre los número pares desde el 1 al 200
         // ^ Declaramos la variable i (incremento)
-        long i;
+        short i;
 
         // ! Pequeña descripción del programa
         System.out.println("Programa que con un bucle for muestre los número pares desde el 1 al 200");
